+++ exonum/src/runtime/rust/service.rs
@@ -95,7 +95,7 @@ pub trait Service: ServiceDispatcher + Debug + 'static {
     /// Note that if service was added within genesis block, it will be activated **immediately** and
     /// thus `after_transactions` will be invoked for such a service after the genesis block creation.
     /// If you aren't interested in the processing of for the genesis block, you can use
-    /// [`CallContext::in_genesis_block`] method and exit early for `true` return value.
+    /// [`CallContext::in_genesis_block`] method and exit early if `true` is returned.
     ///
     /// Also note that invocation of [`blockchain::Schema::height`] will **panic** if invoked within
     /// `after_transactions` of the genesis block. If you are going to process the genesis
@@ -106,7 +106,7 @@ pub trait Service: ServiceDispatcher + Debug + 'static {
     ///
     /// [`blockchain::Schema::height`]: ../../blockchain/schema/struct.Schema.html#method.height
     /// [`blockchain::Schema::height`]: ../../blockchain/schema/struct.Schema.html#method.next_height
-    /// [`CallContext::in_genesis_block`]: call_context/struct.CallContext.html#method.in_genesis_block
+    /// [`CallContext::in_genesis_block`]: struct.CallContext.html#method.in_genesis_block
     fn after_transactions(&self, _context: CallContext<'_>) -> Result<(), ExecutionError> {
         Ok(())
     }