+++ CHANGELOG.md
@@ -27,7 +27,7 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   have been removed. (#1553)
 - `after_transactions` hook is now invoked on the genesis block for the builtin
   services. Note that calling `blockchain::Schema::height` within `after_transactions`
-  hook may cause a panic for a builtin service. (#1619)
+  hook will cause a panic for a builtin service. (#1619)
 
 #### exonum-supervisor
 
@@ -42,8 +42,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   an `ExecutionError` has an expected type, error message and/or location. (#1585)
 - New `blockchain::Schema` method `next_height` was added as a non-panicking
   alternative to `height`. (#1619)
-- New macro `skip_for_genesis` was added for `rust` services to skip service hook
-  execution for the genesis block. (#1619)
+- New method `in_genesis_block` was added to the `CallContext` to check if the service
+  hook is being executed for the genesis block. (#1619)
 
 #### exonum-merkledb